[Liquidity Mining] Code reorg (9) (#208)
* Moving logic into respective modules and hiding accessors

* Making some accessors public for tests

* Increasing budget limit

* Adding a test to add pool reward

* Adding borrow position kind
